+++ NEWS.md
@@ -1,12 +1,15 @@
 # DIZtools NEWS
 
-## Unreleased (2022-03-01)
+## Unreleased (2022-03-03)
 
 #### Fixes
 
 * global option now work for feedback
+#### Docs
 
-Full set of changes: [`v0.0.3...0042835`](https://gitlab.miracum.org/miracum/misc/diztools/compare/v0.0.3...0042835)
+* updated news.md
+
+Full set of changes: [`v0.0.3...fce9ddf`](https://gitlab.miracum.org/miracum/misc/diztools/compare/v0.0.3...fce9ddf)
 
 ## v0.0.3 (2022-02-24)