+++ NEWS.md
@@ -11,8 +11,9 @@
 #### Docs
 
 * updated news.md
+* updated news.md
 
-Full set of changes: [`v0.0.3...a241b68`](https://gitlab.miracum.org/miracum/misc/diztools/compare/v0.0.3...a241b68)
+Full set of changes: [`v0.0.3...9246684`](https://gitlab.miracum.org/miracum/misc/diztools/compare/v0.0.3...9246684)
 
 ## v0.0.3 (2022-02-24)